+++ static/app/views/dashboards/widgetBuilder/components/sectionHeader.tsx
@@ -1,29 +1,42 @@
+import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
 import {Tooltip} from 'sentry/components/tooltip';
+import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 
 interface SectionHeaderProps {
   title: React.ReactNode;
+  optional?: boolean;
   tooltipText?: React.ReactNode;
 }
 
-export function SectionHeader({tooltipText, title}: SectionHeaderProps) {
+export function SectionHeader({tooltipText, title, optional}: SectionHeaderProps) {
   return (
-    <Tooltip
-      title={tooltipText}
-      disabled={!tooltipText}
-      position="right-end"
-      delay={200}
-      isHoverable
-      showUnderline
-    >
-      <StyledHeader>{title}</StyledHeader>
-    </Tooltip>
+    <Fragment>
+      <Tooltip
+        title={tooltipText}
+        disabled={!tooltipText}
+        position="right-end"
+        delay={200}
+        isHoverable
+        showUnderline
+      >
+        <StyledHeader>{title}</StyledHeader>
+      </Tooltip>
+      {optional && <OptionalHeader>{t('(optional)')}</OptionalHeader>}
+    </Fragment>
   );
 }
 
 const StyledHeader = styled('h6')`
   font-size: ${p => p.theme.fontSizeLarge};
+  margin-bottom: ${space(1)};
+`;
+
+const OptionalHeader = styled('h6')`
+  font-size: ${p => p.theme.fontSizeLarge};
+  color: ${p => p.theme.subText};
+  font-weight: ${p => p.theme.fontWeightNormal};
   margin-bottom: ${space(1)};
 `;
